+++ crates/affected/src/affected_tracker.rs
@@ -1,12 +1,14 @@
 use crate::affected::*;
 use moon_common::path::WorkspaceRelativePathBuf;
 use moon_common::{Id, color};
+use moon_config::Input;
 use moon_env_var::GlobalEnvBag;
 use moon_project::Project;
 use moon_task::{Target, Task, TaskOptionRunInCI};
 use moon_workspace_graph::{GraphConnections, WorkspaceGraph};
 use rustc_hash::{FxHashMap, FxHashSet};
 use starbase_utils::fs;
+use starbase_utils::glob::GlobSet;
 use std::fmt;
 use std::sync::Arc;
 use tracing::{debug, trace};
@@ -74,10 +76,11 @@ impl AffectedTracker {
             debug!(
                 env = ?state.env.iter().collect::<Vec<_>>(),
                 files = ?state.files.iter().collect::<Vec<_>>(),
+                projects = ?state.projects.iter().map(|id| id.as_str()).collect::<Vec<_>>(),
                 upstream = ?state.upstream.iter().map(|target| target.as_str()).collect::<Vec<_>>(),
                 downstream = ?state.downstream.iter().map(|target| target.as_str()).collect::<Vec<_>>(),
                 other = state.other,
-                "Task {} is affected by", color::label(&target),
+                "Task {} is affected by", color::id(&target),
             );
 
             affected.tasks.insert(target, state);
@@ -159,6 +162,59 @@ impl AffectedTracker {
                 .find(|file| file.starts_with(&project.source))
                 .map(|file| AffectedBy::TouchedFile(file.to_owned()))
         }
+    }
+
+    pub fn is_project_affected_using_file_group(
+        &self,
+        project: &Project,
+        file_group_id: &Id,
+    ) -> miette::Result<Option<AffectedBy>> {
+        let group = project.get_file_group(file_group_id)?;
+
+        if !group.files.is_empty() {
+            for file in &group.files {
+                if self.touched_files.contains(file) {
+                    return Ok(Some(AffectedBy::TouchedFile(file.to_owned())));
+                }
+            }
+        }
+
+        if !group.globs.is_empty() {
+            return self.is_project_affected_using_globs(project, &group.globs);
+        }
+
+        Ok(None)
+    }
+
+    pub fn is_project_affected_using_globs<G: AsRef<str>>(
+        &self,
+        project: &Project,
+        base_globs: &[G],
+    ) -> miette::Result<Option<AffectedBy>> {
+        let mut globs = vec![];
+
+        // Ensure they are relative from the project root
+        for glob in base_globs {
+            let glob = glob.as_ref();
+
+            if glob.starts_with(project.source.as_str()) {
+                globs.push(glob.to_owned());
+            } else {
+                globs.push(format!("{}/{glob}", project.source));
+            }
+        }
+
+        if !globs.is_empty() {
+            let globset = GlobSet::new(&globs)?;
+
+            for file in &self.touched_files {
+                if globset.matches(file.as_str()) {
+                    return Ok(Some(AffectedBy::TouchedFile(file.to_owned())));
+                }
+            }
+        }
+
+        Ok(None)
     }
 
     pub fn is_project_marked(&self, project: &Project) -> bool {
@@ -332,6 +388,7 @@ impl AffectedTracker {
             return Ok(Some(AffectedBy::AlreadyMarked));
         }
 
+        // Special CI handling
         if self.ci {
             match &task.options.run_in_ci {
                 TaskOptionRunInCI::Always => {
@@ -342,11 +399,12 @@ impl AffectedTracker {
             };
         }
 
-        // inputs: []
+        // Never affected
         if task.state.empty_inputs {
             return Ok(None);
         }
 
+        // By env vars
         if !task.input_env.is_empty() {
             let bag = GlobalEnvBag::instance();
 
@@ -359,30 +417,75 @@ impl AffectedTracker {
             }
         }
 
+        // By files
         let globset = task.create_globset()?;
 
         for file in self.touched_files.iter() {
-            let mut affected = false;
-
-            if let Some(params) = task.input_files.get(file) {
-                affected = match &params.content {
+            let affected = if let Some(params) = task.input_files.get(file) {
+                match &params.content {
                     Some(matcher) => {
-                        let content =
-                            fs::read_file(file.to_logical_path(&self.workspace_graph.root))?;
+                        let abs_file = file.to_logical_path(&self.workspace_graph.root);
 
-                        matcher.is_match(&content)
+                        if abs_file.exists() {
+                            matcher.is_match(&fs::read_file(abs_file)?)
+                        } else {
+                            false
+                        }
                     }
                     None => true,
-                };
-            }
-
-            if !affected {
-                affected = globset.matches(file.as_str());
-            }
+                }
+            } else {
+                globset.matches(file.as_str())
+            };
 
             if affected {
                 return Ok(Some(AffectedBy::TouchedFile(file.to_owned())));
             }
+        }
+
+        // By other inputs
+        let mut has_all_project_sources = false;
+
+        for input in &task.inputs {
+            match input {
+                Input::Project(inner) => {
+                    if has_all_project_sources {
+                        continue;
+                    }
+
+                    let projects = if inner.project == "^" {
+                        has_all_project_sources = true;
+
+                        let parent = self
+                            .workspace_graph
+                            .get_project(task.target.get_project_id()?)?;
+
+                        self.workspace_graph
+                            .get_projects_by_id(parent.dependencies.iter().map(|dep| &dep.id))?
+                    } else {
+                        vec![self.workspace_graph.get_project(&inner.project)?]
+                    };
+
+                    for project in projects {
+                        let affected = if let Some(group_id) = &inner.group {
+                            self.is_project_affected_using_file_group(&project, group_id)?
+                                .is_some()
+                        } else if !inner.filter.is_empty() {
+                            self.is_project_affected_using_globs(&project, &inner.filter)?
+                                .is_some()
+                        } else {
+                            self.is_project_affected(&project).is_some()
+                        };
+
+                        if affected {
+                            return Ok(Some(AffectedBy::UpstreamProject(project.id.clone())));
+                        }
+                    }
+                }
+                _ => {
+                    // Skip
+                }
+            };
         }
 
         Ok(None)