+++ crates/project/src/lib.rs
@@ -1,5 +1,7 @@
 mod project;
+mod project_error;
 
 pub use moon_config::{LayerType, ProjectConfig, StackType};
 pub use moon_file_group::FileGroup;
 pub use project::*;
+pub use project_error::*;