+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Added a new task input type that allows you to depend on changes to a project's files directly,
+  instead of depending on a project's task.
+  - Added URI support: `project://<id>`
+  - Added object support: `project: '<id>'`
+  - Can filter with globs or by file group.
+  - Can reference all project dependencies with `^`.
+- Added a new task input type for referencing file groups within the owning project.
+  - Added URI support: `group://<name>`
+  - Added object support: `group: '<name>'`
+  - This is similar to the `@files`, `@globs`, etc, token functions.
+
 #### ⚙️ Internal
 
 - Updated proto to [v0.53.0](https://github.com/moonrepo/proto/releases/tag/v0.53.0) (from 0.52.3).